--- a/.gitignore
+++ b/.gitignore
@@ -15,3 +15,4 @@
 
 vendor
 .vscode
+.DS_Store