--- a/engine_test.go
+++ b/engine_test.go
@@ -6,7 +6,7 @@ import (
 	"runtime"
 	"testing"
 
-	"github.com/json-iterator/go"
+	jsoniter "github.com/json-iterator/go"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -326,10 +326,10 @@ func BenchmarkRun(b *testing.B) {
 
 			json.Unmarshal(byteValue, &ruleMap)
 
-			e := New()
+			e := NewWithCustomWorkerSize(test.Worker)
 			e.Set(test.Facts, ruleMap.Data, test.RuleFunction)
 			for i := 0; i < b.N; i++ {
-				e.Run(test.Target, test.Worker)
+				e.Run(test.Target, 0)
 			}
 		})
 	}
@@ -435,9 +435,9 @@ func BenchmarkRunFullRemakeEngine(b *testing.B) {
 			json.Unmarshal(byteValue, &ruleMap)
 
 			for i := 0; i < b.N; i++ {
-				e := New()
+				e := NewWithCustomWorkerSize(test.Worker)
 				e.Set(test.Facts, ruleMap.Data, test.RuleFunction)
-				e.Run(test.Target, test.Worker)
+				e.Run(test.Target, 0)
 			}
 		})
 	}