# Changelog
- Implement referenced pooling
- Implement Runtime with pooling

# Summary
- `Run()` with reused engine will perform up to 100% (or more) faster
- `Run()` with new engine (each loop) will perform 15-20% slower

# Benchmark Result

## Before
```
Running tool: /usr/local/bin/go test -benchmem -run=^$ github.com/hooqtv/fished -bench ^(BenchmarkRun|BenchmarkRunFullRemakeEngine)$

goos: darwin
goarch: amd64
pkg: github.com/hooqtv/fished
BenchmarkRun/DefaultWorker-4     	  100000	     20902 ns/op	    1409 B/op	      27 allocs/op
BenchmarkRun/1_Worker-4          	  100000	     12643 ns/op	    1408 B/op	      27 allocs/op
BenchmarkRun/Equal_CPU_Count-4   	  100000	     22430 ns/op	    1409 B/op	      27 allocs/op
BenchmarkRun/Double_CPU_Count-4  	   50000	     29812 ns/op	    1410 B/op	      27 allocs/op
BenchmarkRun/DefaultWorker_w/_rf-4         	  100000	     15393 ns/op	    1121 B/op	      20 allocs/op
BenchmarkRunFullRemakeEngine/DefaultWorker-4         	   20000	     70643 ns/op	   18371 B/op	     302 allocs/op
BenchmarkRunFullRemakeEngine/1_Worker-4              	   20000	     60419 ns/op	   18372 B/op	     302 allocs/op
BenchmarkRunFullRemakeEngine/Equal_CPU_Count-4       	   20000	     78171 ns/op	   18369 B/op	     302 allocs/op
BenchmarkRunFullRemakeEngine/Double_CPU_Count-4      	   20000	     93516 ns/op	   18375 B/op	     302 allocs/op
BenchmarkRunFullRemakeEngine/DefaultWorker_w/_rf-4   	   30000	     39711 ns/op	    7918 B/op	     131 allocs/op
PASS
ok  	github.com/hooqtv/fished	20.416s
Success: Benchmarks passed.
```

## After
```
Running tool: /usr/local/bin/go test -benchmem -run=^$ github.com/hooqtv/fished -bench ^(BenchmarkRun|BenchmarkRunFullRemakeEngine)$

goos: darwin
goarch: amd64
pkg: github.com/hooqtv/fished
BenchmarkRun/DefaultWorker-4     	  200000	     10383 ns/op	     912 B/op	      20 allocs/op
BenchmarkRun/1_Worker-4          	  200000	      7581 ns/op	     912 B/op	      20 allocs/op
BenchmarkRun/Equal_CPU_Count-4   	  200000	     10740 ns/op	     912 B/op	      20 allocs/op
BenchmarkRun/Double_CPU_Count-4  	  200000	     10696 ns/op	     913 B/op	      20 allocs/op
BenchmarkRun/DefaultWorker_w/_rf-4         	  300000	      5724 ns/op	     656 B/op	      13 allocs/op
BenchmarkRunFullRemakeEngine/DefaultWorker-4         	   20000	     93671 ns/op	   22349 B/op	     314 allocs/op
BenchmarkRunFullRemakeEngine/1_Worker-4              	   20000	     76087 ns/op	   21466 B/op	     310 allocs/op
BenchmarkRunFullRemakeEngine/Equal_CPU_Count-4       	   20000	     99513 ns/op	   22976 B/op	     316 allocs/op
BenchmarkRunFullRemakeEngine/Double_CPU_Count-4      	   10000	    100829 ns/op	   25127 B/op	     326 allocs/op
BenchmarkRunFullRemakeEngine/DefaultWorker_w/_rf-4   	   30000	     67783 ns/op	   12460 B/op	     145 allocs/op
PASS
ok  	github.com/hooqtv/fished	25.359s
Success: Benchmarks passed.
```

# Reference
https://www.akshaydeo.com/blog/2017/12/23/How-did-I-improve-latency-by-700-percent-using-syncPool/